requirements.txt

flask
flask-socketio
eventlet          
python-dotenv

numpy<2.0.0
opencv-python-headless
ultralytics
onnx
ncnn

Pillow>=9.0.0
